# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Skew AI — Voice-of-Customer Intelligence Platform
  description: 'Skew AI (skewai): domain-agnostic voice-of-customer intelligence.
    Plug in a Domain Pack and get an AI voice agent, a swarm of triage/RCA agents,
    and the Qubot v2 auditor.


    ## Auth

    When `FRONTLINE_API_KEY` is set, send `X-API-Key` or `Authorization: Bearer`.


    ## Versioning

    Current API version **1** (header `API-Version`). Paths under `/api/...` are v1
    without a URL prefix. Breaking changes will introduce `/api/v2/...`.


    ## Errors

    Errors use RFC 7807-shaped JSON (`type`, `title`, `status`, `detail`) while keeping
    FastAPI''s `detail` string for existing clients.


    ## Pagination

    Collection endpoints accept `limit`, `offset`, and optional `cursor`, and return
    a `pagination` object (`has_more`, `next_offset`, `next_cursor`).'
  version: 2.0.0
paths:
  /api/interactions/start:
    post:
      tags:
      - interactions
      summary: Start Interaction
      description: Create an interaction + return the WS URL + greeting.
      operationId: start_interaction_api_interactions_start_post
      parameters:
      - name: channel
        in: query
        required: false
        schema:
          type: string
          default: web_voice
          title: Channel
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Start Interaction Api Interactions Start Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/interactions/ingest:
    post:
      tags:
      - interactions
      summary: Ingest Interaction
      description: Non-conversational complaint ingest (ticket/webhook path).
      operationId: ingest_interaction_api_interactions_ingest_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - type: object
                additionalProperties: true
              - type: 'null'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Ingest Interaction Api Interactions Ingest Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/interactions:
    get:
      tags:
      - interactions
      summary: List Interactions
      description: List interactions (optionally filter by status). Paginated.
      operationId: list_interactions_api_interactions_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Interactions Api Interactions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/interactions/{interaction_id}:
    get:
      tags:
      - interactions
      summary: Get Interaction
      description: Return the interaction header + turns + actions for replay.
      operationId: get_interaction_api_interactions__interaction_id__get
      parameters:
      - name: interaction_id
        in: path
        required: true
        schema:
          type: string
          title: Interaction Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Interaction Api Interactions  Interaction Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/interactions/{interaction_id}/end:
    post:
      tags:
      - interactions
      summary: End Interaction
      description: End the interaction (customer hung up via REST instead of WS).
      operationId: end_interaction_api_interactions__interaction_id__end_post
      parameters:
      - name: interaction_id
        in: path
        required: true
        schema:
          type: string
          title: Interaction Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response End Interaction Api Interactions  Interaction Id  End
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/interactions/{interaction_id}/takeover:
    post:
      tags:
      - interactions
      summary: Takeover
      description: Supervisor takes over the conversation.
      operationId: takeover_api_interactions__interaction_id__takeover_post
      parameters:
      - name: interaction_id
        in: path
        required: true
        schema:
          type: string
          title: Interaction Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Takeover Api Interactions  Interaction Id  Takeover
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/interactions/{interaction_id}/release:
    post:
      tags:
      - interactions
      summary: Release
      description: Supervisor releases; AI resumes.
      operationId: release_api_interactions__interaction_id__release_post
      parameters:
      - name: interaction_id
        in: path
        required: true
        schema:
          type: string
          title: Interaction Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Release Api Interactions  Interaction Id  Release
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...